diabetics/requirements.txt
absl-py==0.15.0
asgiref==3.4.1
astunparse==1.6.3
backports.entry-points-selectable==1.1.0
cachetools==4.2.4
certifi==2021.10.8
cffi==1.14.6
charset-normalizer==2.0.6
clang==5.0
colorama==0.4.4
cryptography==35.0.0
cycler==0.10.0
defusedxml==0.7.1
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-allauth==0.45.0
django-cors-headers==3.9.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-passwordreset==1.2.1
djangorestframework==3.12.4
filelock==3.0.12
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.34.1
gunicorn==20.1.0
h5py==3.1.0
heroku==0.1.4
idna==3.2
joblib==1.1.0
keras==2.6.0
keras-nightly==2.8.0.dev2021102707
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
lightgbm==3.3.0
Markdown==3.3.4
matplotlib==3.4.3
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.58
opt-einsum==3.3.0
pandas==1.3.4
Pillow==8.3.2
pipenv==2021.5.29
platformdirs==2.2.0
protobuf==3.19.0
psycopg2==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pyparsing==3.0.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==1.0
scipy==1.7.1
six==1.15.0
sqlparse==0.4.1
tb-nightly==2.7.0a20211013
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.5.0
tensorflow-estimator==2.5.0
tensorflow-io-gcs-filesystem==0.21.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021102708
tf-nightly-gpu==2.8.0.dev20211026
threadpoolctl==3.0.0
typing-extensions==3.7.4.3
urllib3==1.26.7
virtualenv==20.7.2
virtualenv-clone==0.5.7
virtualenvwrapper-win==1.2.6
Werkzeug==2.0.2
whitenoise==5.3.0
wrapt==1.12.1
